Test opensearch migration
