+++ docker/elasticsearch/Dockerfile
@@ -1,4 +1,4 @@
-FROM amazon/opendistro-for-elasticsearch:1.13.2
+FROM opensearchproject/opensearch:1.3.2
 
 RUN elasticsearch-plugin install --batch \
   analysis-icu \
