+++ src/test/java/org/folio/search/support/extension/impl/ElasticSearchContainerExtension.java
@@ -9,7 +9,7 @@ import org.testcontainers.images.builder.ImageFromDockerfile;
 
 public class ElasticSearchContainerExtension implements BeforeAllCallback, AfterAllCallback {
 
-  private static final String SPRING_PROPERTY_NAME = "spring.elasticsearch.uris";
+  private static final String SPRING_PROPERTY_NAME = "spring.opensearch.uris";
   private static final String ES_IMAGE_NAME = "opendistro-elasticsearch-it:1.13.2";
   private static final Path ES_DOCKERFILE_PATH = Path.of("docker/elasticsearch/Dockerfile");
   private static final GenericContainer<?> CONTAINER = createContainer();
